requirements.txt
fastapi
uvicorn
python-multipart
pydantic>=2.0
pydantic-settings
pytest
httpx
python-dotenv
python-multipart
requests
pyodbc
pytest-cov
ruff
pre-commit

azure-core
azure-cognitiveservices-speech

python-jose[cryptography]
passlib[bcrypt]

opencensus-ext-azure
opencensus-ext-fastapi
opencensus-ext-logging
opencensus-ext-requests
